+++ CHANGELOG.md
@@ -1,3 +1,5 @@
-# 26/11/2019
+# 04/12/2019   
+- Allow microservices to update the dataset's `errorMessage` field
 
+# 26/11/2019
 - Added `subscribable` boolean filter in GET /datasets endpoint.